refactor(sampling): isolate Pauli preparation
