+++ src/clifft/sampling/kernels.h
@@ -1,49 +1,14 @@
 #pragma once
 
-#include "clifft/sampling/plan.h"
+#include "clifft/sampling/pauli_preparation.h"
 #include "clifft/sampling/state.h"
 
-#include <complex>
 #include <cstdint>
 
 namespace clifft::sampling {
 
-// Portable descriptors and scalar kernel entry points used by CPU execution.
-// ISA-specific selection and implementations stay outside this header.
-
-// Describes a Pauli operation to apply to the state vector. Its masks and
-// precomputed index values identify affected coefficients without referring to
-// the State's storage.
-struct PreparedPauli {
-    uint32_t active_width = 0;
-    uint64_t x = 0;
-    uint64_t z = 0;
-    // Highest set X bit. It selects the pair stride and half-space for a
-    // non-diagonal Pauli, and is zero only when the Pauli is diagonal.
-    uint64_t pairing_bit = 0;
-    std::complex<double> even_phase = {1.0, 0.0};
-
-    [[nodiscard]] bool is_identity() const { return x == 0 && z == 0; }
-    [[nodiscard]] bool is_diagonal() const { return x == 0; }
-};
-
-struct PreparedRotation {
-    PreparedPauli pauli;
-    double cosine = 1.0;
-    double sine = 0.0;
-};
-
-struct PreparedPromotion {
-    double cosine = 1.0;
-    double sine = 0.0;
-};
-
-struct PreparedMeasurement {
-    PreparedPauli pauli;
-    uint32_t pivot = 0;
-    uint64_t output_size = 0;
-    uint64_t z_without_pivot = 0;
-};
+// Portable scalar kernel entry points used by CPU execution. ISA-specific
+// selection and implementations stay outside this header.
 
 struct MeasurementProbabilities {
     double zero = 0.0;
@@ -52,13 +17,6 @@ struct MeasurementProbabilities {
     [[nodiscard]] double total() const { return zero + one; }
     [[nodiscard]] double for_branch(bool branch) const { return branch ? one : zero; }
 };
-
-[[nodiscard]] PreparedPauli prepare_pauli(ActivePauli pauli, uint32_t active_width);
-[[nodiscard]] PreparedRotation prepare_rotation(ActivePauli pauli, uint32_t active_width,
-                                                double half_turns);
-[[nodiscard]] PreparedPromotion prepare_promotion(double half_turns);
-[[nodiscard]] PreparedMeasurement prepare_measurement(ActivePauli pauli, uint32_t active_width,
-                                                      uint32_t pivot);
 
 // Computes <P> on normalized active-coordinate coefficients without mutating
 // the state.
